requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
ofxparse>=0.21
scikit-learn>=1.3.0
pandas>=2.0.0
beancount>=2.3.5
pydantic>=2.4.0
PyYAML>=6.0
prompt_toolkit>=3.0.0
requests>=2.31.0
click>=8.1.0
rapidfuzz>=3.5.0